PP-3812 Add cypress transaction assertions for amounts
PP-3812 Added pact/contract for transactions filtered by from_date

PP-3812 Added cypress tests to check transaction filtering using a 'from' date/time

PP-3812 Added connector pact/contract interactions for filtering transactions using a 'to_date'

PP-3812 Cypress transaction filtering tests now use 'to_date' filtering assertions
